fixed bugs, added parents route

+++ app/serializer.py
@@ -733,7 +733,7 @@ class PromoteSerializer(serializers.ModelSerializer):
         if member is None:
             
             if role == "member":
-                member = CommunityMember.objects.create(community=instance, user_id=user_id)
+                member = CommunityMember.objects.create(community=instance, user_id=user_id).first()
                 member.is_reviewer = False
                 member.is_moderator = False
                 member.is_admin = False
@@ -749,12 +749,13 @@ class PromoteSerializer(serializers.ModelSerializer):
                 raise serializers.ValidationError(detail={"error": "role can't be None"})
             
             elif role == 'reviewer':
-                moderator = Moderator.objects.filter(user_id=user_id, community=instance)
-                article_moderator = ArticleModerator.objects.filter(moderator_id=moderator.id)
-                if article_moderator.exists():
-                    raise serializers.ValidationError(detail={"error": "user is moderator of some articles.Can not perform this operation!!!"})
-                if moderator.exists():
-                    moderator.delete()
+                moderator = Moderator.objects.filter(user_id=user_id, community=instance).first()
+                if moderator is not None:
+                    article_moderator = ArticleModerator.objects.filter(moderator_id=moderator.id)
+                    if article_moderator.exists():
+                        raise serializers.ValidationError(detail={"error": "user is moderator of some articles.Can not perform this operation!!!"})
+                    else :
+                        moderator.delete()
                 OfficialReviewer.objects.create(User_id=user_id, community=instance, Official_Reviewer_name=fake.name())
                 member.is_reviewer = True
                 member.is_moderator = False
@@ -764,11 +765,11 @@ class PromoteSerializer(serializers.ModelSerializer):
                 UserActivity.objects.create(user=self.context['request'].user, action=f'you added {member.user.username} to {instance.Community_name} as reviewer')
                 
             elif role == 'moderator':
-                reviewer = OfficialReviewer.objects.filter(User_id=user_id, community=instance)
-                article_reviewer = ArticleReviewer.objects.filter(officialreviewer_id=reviewer.id)
-                if article_reviewer.exists():
-                    raise serializers.ValidationError(detail={"error": "user is reviewer of some articles.Can not perform this operation!!!"})
-                if reviewer.exists():
+                reviewer = OfficialReviewer.objects.filter(User_id=user_id, community=instance).first()
+                if reviewer is not None:
+                    article_reviewer = ArticleReviewer.objects.filter(officialreviewer_id=reviewer.id)
+                    if article_reviewer.exists():
+                        raise serializers.ValidationError(detail={"error": "user is reviewer of some articles.Can not perform this operation!!!"})
                     reviewer.delete()
                 Moderator.objects.create(user_id=user_id, community=instance)
                 member.is_moderator = True
@@ -780,11 +781,18 @@ class PromoteSerializer(serializers.ModelSerializer):
                 
             elif role == 'admin':
                 reviewer = OfficialReviewer.objects.filter(User_id=user_id, community=instance).first()
-                if reviewer is not None:
-                    reviewer.delete()
                 moderator = Moderator.objects.filter(user_id=user_id, community=instance).first()
+                if reviewer is not None:
+                    article_reviewer = ArticleReviewer.objects.filter(officialreviewer_id=reviewer.id)
+                    if article_reviewer.exists():
+                        raise serializers.ValidationError(detail={"error": "user is reviewer of some articles.Can not perform this operation!!!"})
+                    reviewer.delete()
                 if moderator is not None:
-                    moderator.delete()
+                    article_moderator = ArticleModerator.objects.filter(moderator_id=moderator.id)
+                    if article_moderator.exists():
+                        raise serializers.ValidationError(detail={"error": "user is moderator of some articles.Can not perform this operation!!!"})
+                    else :
+                        moderator.delete()
                 member.is_moderator = False
                 member.is_reviewer = False
                 member.is_admin = True
@@ -793,14 +801,19 @@ class PromoteSerializer(serializers.ModelSerializer):
                 UserActivity.objects.create(user=self.context['request'].user, action=f'you added {member.user.username} to {instance.Community_name} as admin')
                                 
             elif role == 'member':
-                reviewer = OfficialReviewer.objects.filter(User_id=user_id, community=instance)
+                reviewer = OfficialReviewer.objects.filter(User_id=user_id, community=instance).first()
+                moderator = Moderator.objects.filter(user_id=user_id, community=instance).first()
                 if reviewer.exists():
+                    article_reviewer = ArticleReviewer.objects.filter(officialreviewer_id=reviewer.id)
+                    if article_reviewer.exists():
+                        raise serializers.ValidationError(detail={"error": "user is reviewer of some articles.Can not perform this operation!!!"})
                     reviewer.delete()
-                
-                moderator = Moderator.objects.filter(user_id=user_id, community=instance)
                 if moderator.exists():
-                    moderator.delete()
-                    
+                    article_moderator = ArticleModerator.objects.filter(moderator_id=moderator.id)
+                    if article_moderator.exists():
+                        raise serializers.ValidationError(detail={"error": "user is moderator of some articles.Can not perform this operation!!!"})
+                    else :
+                        moderator.delete()
                 member.is_reviewer = False
                 member.is_moderator = False
                 member.save()
@@ -1629,7 +1642,6 @@ class CommentlistSerializer(serializers.ModelSerializer):
         serializer = CommentSerializer(comment,many=True, context={'request': self.context['request']})
         return serializer.data
 
-
 class CommentSerializer(serializers.ModelSerializer):
     user = serializers.SerializerMethodField(read_only=True)
     rank = serializers.SerializerMethodField(read_only=True)
@@ -1768,6 +1780,13 @@ class CommentSerializer(serializers.ModelSerializer):
         comment = CommentBase.objects.filter(version=obj)
         serializer = CommentSerializer(comment,many=True, context={"request": self.context['request']})
         return serializer.data
+    
+class CommentParentSerializer(serializers.ModelSerializer):
+
+    class Meta:
+        model = CommentBase
+        fields = ['id']
+        read_only_fields = ['id']
 
 
 class CommentCreateSerializer(serializers.ModelSerializer):
@@ -1808,19 +1827,19 @@ class CommentCreateSerializer(serializers.ModelSerializer):
 
         if validated_data["parent_comment"]:
             member = CommentBase.objects.filter(id=validated_data['parent_comment'].id).first()
-            notification = Notification.objects.create(user=member.User, message=f'{handler.handle_name} replied to your comment on {member.article.article_name} ', link=f'/article/{member.article.id}/{instance.id}')
-            notification.save()
-            send_mail(f"somebody replied to your comment",f"{handler.handle_name} have made a replied to your comment.checkout this {parse(config('CLIENT_URL'))}/article/{member.article.id}/{instance.id}", settings.EMAIL_HOST_USER,[member.User.email], fail_silently=False)
+            # notification = Notification.objects.create(user=member.User, message=f'{handler.handle_name} replied to your comment on an article', link=f'/article/{member.article.id}/{instance.id}')
+            # notification.save()
+            send_mail(f"somebody replied to your comment",f"{handler.handle_name} have made a replied to your comment.\n {settings.BASE_URL}/article/{member.article.id}/{instance.id}", settings.EMAIL_HOST_USER,[member.User.email], fail_silently=False)
 
         if validated_data["Type"] == "review" or validated_data["Type"] == "decision":
             emails = [author.User.email for author in authors ]
             for author in authors:
-                notification = Notification.objects.create(user=author.User, message=f'{handler.handle_name} has added a {validated_data["Type"]} to your article: {instance.article.article_name} ', link=f'/article/{member.article.id}/{instance.id}')
-                notification.save()
-            send_mail(f"A new {validated_data['Type']} is added ",f"{handler.handle_name} has added a {validated_data['Type']} to your article: {instance.article.article_name}. checkout this {parse(config('CLIENT_URL'))}/article/{member.article.id}/{instance.id}", settings.EMAIL_HOST_USER,emails, fail_silently=False)
+                # notification = Notification.objects.create(user=author.User, message=f'{handler.handle_name} has added a {validated_data["Type"]} to your article', link=f'/article/{instance.article.id}/{instance.id}')
+                # notification.save()
+                send_mail(f"A new {validated_data['Type']} is added ",f"{handler.handle_name} has added a {validated_data['Type']} to your article: {instance.article.article_name}. checkout this {settings.BASE_URL}/article/{instance.article.id}/{instance.id}", settings.EMAIL_HOST_USER,emails, fail_silently=False)
 
             
-        send_mail(f"you have made {instance.Type}",f"You have made a {instance.Type} on {instance.article.article_name}. checkout this {parse(config('CLIENT_URL'))}/article/{member.article.id}/{instance.id}", settings.EMAIL_HOST_USER,[instance.User.email], fail_silently=False)
+        send_mail(f"you have made {instance.Type}",f"You have made a {instance.Type} on {instance.article.article_name}. checkout this {settings.BASE_URL}/article/{instance.article.id}/{instance.id}", settings.EMAIL_HOST_USER,[instance.User.email], fail_silently=False)
         UserActivity.objects.create(user=self.context['request'].user, action=f"You have made a {instance.Type} on {instance.article.article_name}")
 
         return instance    
@@ -1903,7 +1922,6 @@ class ArticleChatCreateSerializer(serializers.ModelSerializer):
         :return: The instance of the created object is being returned.
         """
         article_id = validated_data.get("article")
-        print(article_id)
         validated_data.pop("article")
         article = Article.objects.filter(article_name=article_id).first()
         channel = f"{article.id}"
@@ -2415,7 +2433,7 @@ class SocialPostCommentCreateSerializer(serializers.ModelSerializer):
         instance = self.Meta.model.objects.create(**validated_data, user=self.context['request'].user)
         instance.save()
         if instance.parent_comment is None:
-            notification = Notification.objects.create(user=instance.User, message=f'someone replied to your post on {parse(config("CLIENT_URL"))} ', link=f'/article/{member.article.id}/{instance.id}')
+            notification = Notification.objects.create(user=instance.user, message=f'someone replied to your post', link=f'/post/{instance.post.id}/{instance.id}')
             notification.save()
         return instance
     
@@ -2864,94 +2882,4 @@ class MessageListSerializer(serializers.ModelSerializer):
         """
         count = PersonalMessage.objects.filter(Q(sender=self.context['request'].user) | Q(receiver=self.context['request'].user), is_read=False).count()
         return count
-    
-
-class GroupSerializer(serializers.ModelSerializer):
-
-    class Meta:
-        model = Group
-        fields = ["id", "title", "user", "members" ]
-        read_only_fields = ["id"]
-    
-    
-class GroupUpdateSerializer(serializers.ModelSerializer):
-
-    class Meta:
-        model = Group
-        fields = ["id", "title", "user", "members" ]
-        read_only_fields = ["id", "user", "title"]
-    
-    def update(self, instance, validated_data):
-
-        instance.title = validated_data.get('title', instance.title)
-        instance.save()
-        return instance
-
-class GroupCommentSerializer(serializers.ModelSerializer):
-
-    class Meta:
-        model = GroupComment
-        fields = ["id", "user", "group", "comment", "created_at"]
-        read_only_fields = ["id", "user", "group", "created_at"]
-
-    def create(self, validated_data):
-
-        instance = self.Meta.model.objects.create(**validated_data, user=self.context['request'].user)
-        instance.save()
-        return instance
-    
-class GroupCommentUpdateSerializer(serializers.ModelSerializer):
-
-    class Meta:
-        model = GroupComment
-        fields = ["id", "user", "group", "comment", "created_at"]
-        read_only_fields = ["id", "user", "group", "created_at"]
-
-    def update(self, instance, validated_data):
-
-        instance.comment = validated_data.get('comment', instance.comment)
-        instance.save()
-        return instance
-
-
-class GroupMemberSerializer(serializers.ModelSerializer):
-
-    class Meta:
-        model = GroupMember
-        fields = ["id", "user", "group", "is_admin"]
-        read_only_fields = ["id", "user", "group"]
-    
-    def create(self, validated_data):
-            
-        instance = self.Meta.model.objects.create(**validated_data, user=self.context['request'].user)
-        instance.save()
-        return instance
-    
-class GroupMetaSerializer(serializers.ModelSerializer):
-
-    class Meta:
-        model = GroupMeta
-        fields = ["id", "group", "article"]
-        read_only_fields = ["id", "group"]
-    
-    def create(self, validated_data):
-            
-        instance = self.Meta.model.objects.create(**validated_data, user=self.context['request'].user)
-        instance.save()
-        return instance
-
-class GroupRequestSerializer(serializers.ModelSerializer):
-
-    class Meta:
-        model = GroupRequest
-        fields = ["id", "user", "group", "status"]
-        read_only_fields = ["id", "user", "group"]
-    
-    def create(self, validated_data):
-            
-        instance = self.Meta.model.objects.create(**validated_data, user=self.context['request'].user)
-        instance.save()
-        return instance
-
-
